test(LottoFixture.java) : fixture 패키지 별도 분리
