+++ src/test/java/lotto/lotto/domain/fixture/LottoMoneyFixture.java
@@ -1,4 +1,6 @@
-package lotto.lotto.domain;
+package lotto.lotto.domain.fixture;
+
+import lotto.lotto.domain.LottoMoney;
 
 public class LottoMoneyFixture {
     public static LottoMoney 로또구매금액_10_000() {
