+++ src/test/java/lotto/winning/domain/ReturnRateTest.java
@@ -3,8 +3,8 @@ package lotto.winning.domain;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import static lotto.lotto.domain.LottoMoneyFixture.로또구매금액_10_000;
-import static lotto.winning.domain.TotalWinningMoneyFixture.당첨금액_5000;
+import static lotto.lotto.domain.fixture.LottoMoneyFixture.로또구매금액_10_000;
+import static lotto.winning.fixture.TotalWinningMoneyFixture.당첨금액_5000;
 import static org.assertj.core.api.Assertions.assertThat;
 
 @DisplayName("수익률")
