+++ src/test/java/lotto/winning/domain/TotalWinningMoneyTest.java
@@ -1,10 +1,9 @@
 package lotto.winning.domain;
 
-import common.vo.Count;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import static lotto.winning.domain.MatchCountFixture.*;
+import static lotto.winning.fixture.MatchCountFixture.*;
 import static org.assertj.core.api.Assertions.assertThat;
 
 @DisplayName("당첨 금액 총합")
